+++ site/default.nix
@@ -2,12 +2,11 @@
   site ? null,
   pkgs,
 }:
-with pkgs;
-stdenvNoCC.mkDerivation {
+pkgs.stdenvNoCC.mkDerivation {
   pname = "website";
   version = "stable";
   src = ./.;
-  buildInputs = [ zola ];
+  buildInputs = [ pkgs.zola ];
   installPhase = "cp -r public $out";
   buildPhase = "zola build ${if (site != null) then "--base-url " + site else ""}";
 }