+++ site/git/repos.nix
@@ -1,0 +1,9 @@
+[
+  {
+    name = "matiru";
+    url = "https://github.com/maydayv7/matiru";
+    description = "Blockchain-Based Supply Chain Transparency for Agricultural Produce";
+    owner = "V7";
+    logo = "";
+  }
+]